+++ CHANGES.md
@@ -11,6 +11,9 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 
 ## [Unreleased]
 
+### Added
+- `KtLintStep` now supports additional generated rule-set JARs whose classpath is supplied at execution time. ([#2999](https://github.com/diffplug/spotless/pull/2999), fixes [#1901](https://github.com/diffplug/spotless/issues/1901))
+
 ## [4.8.0] - 2026-06-29
 ### Added
 - Add support for custom string format for license header copyright year via `yearStringFormat()`. ([#2965](https://github.com/diffplug/spotless/pull/2965))